requirements.txt
python-telegram-bot==22.7
requests==2.32.3
python-dotenv==1.0.1